+++ crates/tad-gui/src/song_tab.rs
@@ -693,7 +693,7 @@ impl State {
                 let text = format!(
                     "MML compiled successfully: {} bytes (+{} echo buffer bytes)\n\nDuration: {}\n{}",
                     sd.data().len(),
-                    sd.metadata().echo_buffer.edl.buffer_size(),
+                    sd.metadata().echo_buffer.buffer_size(),
                     song_duration_string(sd.duration()),
                     MmlTickCountTable(&sd),
                 );